requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
Pillow==10.2.0
python-dotenv==1.0.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1 